feat: 플레이어의 이름이 all인 경우 예외 발생기능 추가 + Name -> PlayerName 클래스이름 변경
