+++ src/main/java/domain/PlayerName.java
@@ -1,0 +1,43 @@
+package domain;
+
+public class PlayerName {
+    public static final int MAX_LENGTH = 5;
+    private static final String BLANK_NAME_MESSAGE = "빈 문자열 입니다.";
+    private static final String INVALID_NAME_LENGTH_MESSAGE = "이름 길이는 %d자를 넘길 수 없습니다.";
+    private static final String IMPOSSIBLE_NAME_NOTICE = "불가능한 이름입니다.";
+    private static final String RESERVED_NAME = "all";
+
+
+    private final String name;
+
+    public PlayerName(String name) {
+        this.name = name;
+        validateNameLength(name);
+        validateNameBlank(name);
+        validateName(name);
+    }
+
+    private void validateNameBlank(String name) {
+        if (name.isBlank()) {
+            throw new IllegalArgumentException(BLANK_NAME_MESSAGE);
+        }
+    }
+
+    private void validateNameLength(String name) {
+        if (name.length() > MAX_LENGTH) {
+            throw new IllegalArgumentException(
+                    String.format(INVALID_NAME_LENGTH_MESSAGE, MAX_LENGTH));
+        }
+    }
+
+    private void validateName(String name) {
+        if (name.equals(RESERVED_NAME)) {
+            throw new IllegalArgumentException(IMPOSSIBLE_NAME_NOTICE);
+        }
+    }
+
+    public String getName() {
+        return name;
+    }
+
+}
